refactor(pfb): extract field formatting helpers into useFieldFormat
